Move tracing related methods to tracing pkg
