+++ v2/sender.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
-	"go.opentelemetry.io/otel/propagation"
 )
 
 const (
@@ -104,15 +103,6 @@ func SetMessageTTL(ttl time.Duration) func(msg *azservicebus.Message) error {
 	}
 }
 
-// SetMessageTrace inject the spanContext inside the input context into a service bus message
-func SetMessageTrace(ctx context.Context) func(msg *azservicebus.Message) error {
-	return func(msg *azservicebus.Message) error {
-		propogator := propagation.TraceContext{}
-		propogator.Inject(ctx, carrierAdapter(msg))
-		return nil
-	}
-}
-
 func getMessageType(mb MessageBody) string {
 	var msgType string
 	vo := reflect.ValueOf(mb)
@@ -123,36 +113,4 @@ func getMessageType(mb MessageBody) string {
 	}
 
 	return msgType
-}
-
-type messageWrapper struct {
-	message *azservicebus.Message
-}
-
-// carrierAdapter wraps a servicebus Message so that it implements the TextMapCarrier interface
-func carrierAdapter(message *azservicebus.Message) propagation.TextMapCarrier {
-	return &messageWrapper{message: message}
-}
-
-func (mw *messageWrapper) Set(key string, value string) {
-	if mw.message.ApplicationProperties == nil {
-		mw.message.ApplicationProperties = make(map[string]interface{})
-	}
-	mw.message.ApplicationProperties[key] = value
-}
-
-func (mw *messageWrapper) Get(key string) string {
-	if mw.message.ApplicationProperties == nil || mw.message.ApplicationProperties[key] == nil {
-		return ""
-	}
-
-	return mw.message.ApplicationProperties[key].(string)
-}
-
-func (mw *messageWrapper) Keys() []string {
-	keys := make([]string, 0, len(mw.message.ApplicationProperties))
-	for k := range mw.message.ApplicationProperties {
-		keys = append(keys, k)
-	}
-	return keys
 }
