+++ v2/tracing/tracing_test.go
@@ -2,6 +2,7 @@ package tracing
 
 import (
 	"context"
+	"reflect"
 	"testing"
 	"time"
 
@@ -10,6 +11,7 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/propagation"
 	tracesdk "go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/trace"
 )
 
 func Test_TracingMiddleware(t *testing.T) {
@@ -56,7 +58,7 @@ func Test_TracingMiddleware(t *testing.T) {
 			g := NewGomegaWithT(t)
 			if tc.carryRemoteSpan {
 				propogator := propagation.TraceContext{}
-				propogator.Inject(remoteCtx, carrierAdapter(tc.message))
+				propogator.Inject(remoteCtx, carrierAdapterForReceivedMsg(tc.message))
 				_, remoteSpan := getRemoteParentSpan(context.TODO(), tc.message)
 				g.Expect(remoteSpan.SpanContext().IsValid()).To(BeTrue())
 				g.Expect(remoteSpan.SpanContext().IsRemote()).To(BeTrue())
@@ -65,6 +67,37 @@ func Test_TracingMiddleware(t *testing.T) {
 			_, span := applyTracingMiddleWare(context.TODO(), tc.message)
 			g.Expect(span.SpanContext().IsValid()).To(BeTrue())
 		})
+	}
+}
+
+func TestHandlers_SetMessageTrace(t *testing.T) {
+	blankMsg := &azservicebus.Message{}
+	tp := tracesdk.NewTracerProvider(tracesdk.WithSampler(tracesdk.AlwaysSample()))
+
+	// fake a remote span
+	remoteCtx, remoteSpan := tp.Tracer("test-tracer").Start(context.Background(), "remote-span")
+	remoteSpan.End()
+
+	// Inject the trace context into the message
+	handler := PropagateFromContext(remoteCtx)
+	if err := handler(blankMsg); err != nil {
+		t.Errorf("Unexpected error in set trace carrier test: %s", err)
+	}
+
+	if blankMsg.ApplicationProperties == nil {
+		t.Errorf("for trace carrier expected application properties to be set")
+	}
+
+	propogator := propagation.TraceContext{}
+	ctx := propogator.Extract(context.TODO(), carrierAdapterForMsg(blankMsg))
+	extractedSpan := trace.SpanFromContext(ctx)
+
+	if !extractedSpan.SpanContext().IsValid() || !extractedSpan.SpanContext().IsRemote() {
+		t.Errorf("expected extracted span to be valid and remote")
+	}
+
+	if reflect.DeepEqual(extractedSpan, remoteSpan) {
+		t.Errorf("expected extracted span to be the same with remote span")
 	}
 }
